+++ integrated-scouting-apps/src/index.tsx
@@ -1,15 +1,15 @@
-import React from "react";
+import React from 'react';
 import ReactDOM from 'react-dom/client';
-import HomePage from './routes/home';
-import ScoutingApp from './routes/scoutingapp';
-import MatchScout from './routes/match';
+import HomePage from './routes/homeRouter';
+import ScoutingApp from './routes/scoutingAppRouter';
+import MatchScout from './routes/matchScout';
 import DTF from './routes/dtf';
-import DTFTeams from './routes/dtfteams';
-import StrategicScout from './routes/strategic';
-import Lookup from './routes/lookup';
+import DTFTeams from './routes/dtfTeams';
+import StrategicScout from './routes/strategicScout';
+import Lookup from './routes/lookupRouter';
 import StrategicLookup from './routes/strategicLookup';
 import PitLookup from './routes/pitLookup';
-import PitScout from './routes/pit';
+import PitScout from './routes/pitScout';
 import DataLookup from './routes/matchLookup';
 import TeamData from './routes/matchData';
 import Picklist from './routes/picklist';
