+++ integrated-scouting-apps/src/parts/chart.tsx
