+++ integrated-scouting-apps/src/parts/strategicTabs.tsx
