+++ integrated-scouting-apps/src/routes/matchLookup.tsx
@@ -2,9 +2,9 @@ import '../public/stylesheets/style.css';
 import '../public/stylesheets/lookup.css';
 import { useEffect, useState } from 'react';
 import { Input, Form, InputNumber } from 'antd';
-import Header from './parts/header';
-import { getAllTeams, getDivisionsList } from './utils/tbaRequest';
-import { Select } from './parts/formItems';
+import Header from '../parts/header';
+import { getAllTeams, getDivisionsList } from '../utils/tbaRequest';
+import { Select } from '../parts/formItems';
 
 function TeamData(props: any) {
   const DEFAULT_MATCH_EVENT = process.env.REACT_APP_EVENTNAME || "";
